secure file io example
I found this code to be slightly insecure. So I added some cases to keep if save from likely attacks.

The first check is to handle the case for when the file is not found which is a good practice, also would be good to let the user know that the file was not found. The second check keeps your example program from being closed before returning to the caller.